fixed a few integration tests
