+++ src/integrationTest/java/uk/gov/hmcts/reform/prl/util/ServiceAuthenticationGenerator.java
@@ -11,6 +11,7 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.BEARER;
 
 @TestPropertySource("classpath:application.yaml")
 @Service
@@ -40,5 +41,21 @@ public class ServiceAuthenticationGenerator {
         assertThat(response.getStatusCode()).isEqualTo(200);
 
         return response.getBody().asString();
+    }
+
+    public String generateTokenForCcd() {
+        final Response response = RestAssured
+            .given()
+            .relaxedHTTPSValidation()
+            .baseUri(s2sUrl)
+            .header(CONTENT_TYPE, APPLICATION_JSON_VALUE)
+            .body(Map.of("microservice", "ccd_data"))
+            .when()
+            .post("/testing-support/lease")
+            .andReturn();
+
+        assertThat(response.getStatusCode()).isEqualTo(200);
+
+        return BEARER + response.getBody().asString();
     }
 }
